worker/requirements.txt
flask==3.1.0
python-evtx==0.8.0
yara-python==4.5.1
pefile==2024.8.26
python-registry==1.3.1
waitress==3.0.2
zstandard==0.23.0
